requirement.txt
opencv-python
requests
selenium
webdriver_manager
zai-sdk